Remove swift:4.0, only swift:4.1
