--- core/swift40Action/CHANGELOG.md
@@ -1,20 +1,0 @@
-# Apache OpenWhisk Swift 4.0 Runtime Container
-
-## 1.0.1
-Changes:
-  - Rename image name to `openwhisk/action-swift-v4.0`
-  - Rename kind to `swift:4.0`
-
-Swift runtime version: [4.0.3](https://github.com/IBM-Swift/swift-ubuntu-docker/blob/4a821e2db45f7839789f0f516863be8295142ab1/swift-development/Dockerfile)
-
-
-
-## 1.0.0
-Initial Swift 4.0 image
-  - Image name: `openwhisk/action-swift-v4`
-  - Kind is: `swift:4`
-
-Swift runtime version: [4.0.3](https://github.com/IBM-Swift/swift-ubuntu-docker/blob/4a821e2db45f7839789f0f516863be8295142ab1/swift-development/Dockerfile)
-
-Packages included:
-  - No packages included, use Package.swift and pre-compile action.
