+++ core/swift41Action/build.gradle
@@ -1,36 +1,2 @@
 ext.dockerImageName = 'action-swift-v4.1'
 apply from: '../../gradle/docker.gradle'
-
-distDocker.dependsOn 'copyEpilogue'
-distDocker.dependsOn 'copySwiftRunner'
-distDocker.dependsOn 'copyWhisk'
-distDocker.dependsOn 'copyPackageSwift'
-distDocker.finalizedBy('cleanup')
-
-
-task copyEpilogue(type: Copy) {
-    from '../swift40Action/epilogue.swift'
-    into '.'
-}
-
-task copySwiftRunner(type: Copy) {
-    from '../swift40Action/swift4runner.py'
-    into '.'
-}
-
-task copyWhisk(type: Copy) {
-    from '../swift40Action/spm-build/_Whisk.swift'
-    into './spm-build'
-}
-
-task copyPackageSwift(type: Copy) {
-    from '../swift40Action/spm-build/Package.swift'
-    into './spm-build'
-}
-
-task cleanup(type: Delete) {
-    delete 'epilogue.swift'
-    delete 'swift4runner.py'
-    delete 'spm-build/_Whisk.swift'
-    delete 'spm-build/Package.swift'
-}
