+++ settings.gradle
@@ -2,7 +2,6 @@ include 'tests'
 
 include 'core:swift3.1.1Action'
 
-include 'core:swift40Action'
 include 'core:swift41Action'
 
 rootProject.name = 'runtime-swift'
