+++ tools/travis/publish.sh
@@ -32,8 +32,6 @@ IMAGE_TAG=$3
 
 if [ ${RUNTIME_VERSION} == "3.1.1" ]; then
   RUNTIME="swift3.1.1Action"
-elif [ ${RUNTIME_VERSION} == "4.0" ]; then
-  RUNTIME="swift40Action"
 elif [ ${RUNTIME_VERSION} == "4.1" ]; then
   RUNTIME="swift41Action"
 fi
